Rename `getTokens()` to `completeLogIn()` and update
Do not return tokens. Do not decode access token.
For errors, throw a new error object, to reject the promise, returning an Error object. Returning an object with an `error` string was a remnant from copying code over from the happy body app, where the error string was used in a view.