Add a connect error callback
Allow a callback to be passed, in order to be able to do something when a data API (Socket.IO) connect error happens. For example, when a token is unauthorized or expired.

This is similar to reacting to a 401 with a http API, for an unauthorized call. A common use case is for an app is to log a user out (e.g. deleting tokens and reloading).