Add `isLoggingIn()` utility method
This avoids an annoying and potentially evasive bug which invalidates a login request.

A login button could be rendered at a redirect URI view, which would previously set new PKCE values in localstorage invalidating the login request.

Also useful if a login redirect URI is not used solely to complete login, e.g. an app's home page, to check when `completeLogin()` needs to be called.