Stop invalid login request asap
If a login request is made from an origin that does not make the origin of the loginRedirectUri as configured in the SDK instance, it cannot succeed. Don't bother to open the pop-up or go any further. Show origins in the error to quickly highlight the cause of the error to a developer.

e.g. locally login from `127.0.0.1:8080`, but `localhost:8080` is configured (silly, but frustrating challenging to debug without good errors)
Or a typo was made, or environment config is out of date...