Fix pop-up message event listener clean-up
Previously login was inconsistent is logging in and out. It would fail regularly because of different Window objects where the event source or login window reference or both were incorrectly set.

Make the event handler a class member and memoize in the constructor for consistent clean-up.

This does mean that propoating errors in the event handler is a little more complex, because of the introduction of `resolve` and `reject` members, allowing them to be set from the `login` context so they're available to the event handler.